Extract day route's recurrence-expansion logic into shared util/meetingOccurrences.ts
